requirements.txt
streamlit==1.28.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
Pillow==10.1.0
requests==2.31.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
deep-translator==1.11.4
python-dotenv==1.0.0
groq==0.4.2
pytesseract==0.3.10
pandas>=2.0.0
watchdog>=3.0.0